==> 03-cleaning.Rmd <==
# Data transformation

```{r}
library(tidyverse)
library(readxl)
library(quantmod)
```

In this dataset, many columns are pre-processed by the authors, so our work is to choose the variables we need and change the names and orders of factor levels.

We cleaned the dataset in the following steps:

- Select `AppStore`, `PriceSensitive`, `WhyDownload`, `Currency` & demographic columns (`Gender`, `Age`, `Nationality`, `Education`, `Occupation`, etc.)

- Mutate character variables to factors

- Rename the levels with readable information

- Reorder the levels  for future visualization

```{r}
# import data
app_data = read_excel("data/mobile_app_user_dataset_1.xlsx")
```

```{r}
# choose AppStore, PriceSensitive, WhyDownload & demographic columns
app_user = app_data[-1,] %>%
  select(ID, Q4, Q9_8, Q12.1_2_TEXT, Q16, Q17, Q19, Q23, Q24, Q27) %>%
  rename(AppStore = Q4,
         PriceSensitive = Q9_8,
         WhyDownload = Q12.1_2_TEXT,
         Gender = Q16,
         Age = Q17,
         Nationality = Q19,
         Education = Q23,
         YearsEdu = Q24,
         Occupation = Q27)
```

```{r}
# rename Price Sensitive levels
app_user = app_user %>%
  mutate(PriceSensitive = as.factor(PriceSensitive)) %>%
  mutate(PriceSensitive = fct_recode(
    PriceSensitive, Yes = "1"))
```

```{r}
# rename gender levels
app_user = app_user %>%
  mutate(Gender = as.factor(Gender)) %>%
  mutate(Gender = fct_recode(
    Gender, Male = "1", Female = "2"))
```

```{r}
# rename AppStore levels
app_user = app_user %>%
  mutate(
    AppStore = fct_recode(
      AppStore,
      iOS = "1",
      Blackberry = "2",
      Android = "3",
      Nokia = "4",
      Samsung = "5",
      Windows = "6",
      None = "7",
      `Not sure` = "8",
      Other = "9",
      `Do not use apps` = "10"
    )
  ) %>%
  mutate(AppStore = as.factor(AppStore))
```

```{r}
# rename nationality levels
app_user = app_user %>%
  mutate(Nationality = as.factor(Nationality)) %>%
  mutate(
    Nationality = fct_recode(
      Nationality,
      American = "1",
      Australian = "2",
      Brazillian = "3",
      British = "4",
      Canadian = "5",
      Chinese = "6",
      French = "7",
      German = "8",
      Indian = "9",
      Italian = "10",
      Japanese = "11",
      Mexican = "12",
      Russian = "13",
      SouthKorean = "14",
      Spanish = "15",
      Other = "16"
    ))
```

```{r}
# rename occupation levels
app_user = app_user %>%
  mutate(Occupation = as.factor(Occupation)) %>%
  mutate(
    Occupation = fct_recode(
      Occupation,
      `Mgmt` = "1",
      `Biz & Fin Ops` = "2",
      `Comp & Math` = "3",
      `Arch & Eng` = "4",
      `Life, Phys & Soc Sci` = "5",
      `Cmty & Soc Svcs` = "6",
      Legal = "7",
      `Edu, Trng & Lib` = "8",
      `Arts, Entmt & Sports` = "9",
      `Healthcare Pract & Tech` = "10",
      `Healthcare Sup` = "11",
      `Protective Svcs` = "12",
      `Food Prep & Srv` = "13",
      `Bldg & Grnd Clng` = "14",
      `Prsn Care & Svcs` = "15",
      Sales = "16",
      `Off & Admin Sup` = "17",
      AFF = "18", # Agriculture Forestry & Fishing
      `Constrn & Ext` = "19",
      `Inst, Maint & Repair` = "20",
      Production = "21",
      Transport = "22",
      Military = "23",
      Student = "24",
      Other = "25"
    )
  )
```


```{r}
# reasons why they pay
# multiple choices; won't merge with the main table
WhyPay = app_data[-1,] %>%
  select(ID, starts_with("Q11")) %>%
  rename(
    `I do not pay for apps` = Q11_1,
    `To remove ad` = Q11_2,
    `App is on sale` = Q11_3,
    `Subscription to free content` = Q11_4,
    `Pay for features` = Q11_5,
    `No similar free app` = Q11_6,
    `Better quality in general` = Q11_11,
    `More features in general` = Q11_12,
    `For additional features or content` = Q11_7,
    `Subscription to paid content` = Q11_8,
    `Better quality than free apps`  = Q11_9,
    Other = Q11_10
  )

# merge with price-sensitive users
temp = app_user %>% select(ID, PriceSensitive)
WhyPay = left_join(WhyPay, temp)
```

```{r}
# app categories 
# multiple choices; won't merge with the main table
app_type = app_data[-1, ] %>%
  select(ID, starts_with("Q15")) %>%
  rename(
    Navigation = Q15_1,
    Business = Q15_2,
    Catalogues = Q15_3,
    Travel = Q15_4,
    Books = Q15_5,
    PhotoVideo = Q15_6,
    Lifestyle = Q15_7,
    Entertainment = Q15_8,
    Finance = Q15_9,
    News = Q15_10,
    Health = Q15_11,
    Games = Q15_12,
    Food = Q15_13,
    Education = Q15_14,
    Medical = Q15_15,
    SocialNetworking = Q15_16,
    Reference = Q15_17,
    Sports = Q15_18,
    Utilities = Q15_19,
    Weather = Q15_20,
    Productivity = Q15_21,
    Music = Q15_22,
    Other = Q15_23
  )
```


```{r}
# currency
currency = app_data[-1, ] %>%
  select(ID, starts_with("Q29")) %>%
  rename(
    AUD = Q29.1,
    BRL = Q29.2,
    GBP = Q29.3,
    CAD = Q29.4,
    CNY = Q29.5,
    EUR = Q29.6,
    INR = Q29.7,
    JPY = Q29.8,
    MXN = Q29.9,
    RUB = Q29.10,
    KRW = Q29.11,
    USD = Q29.12,
    Other = Q29.13
  )
```

```{r}
# drop NAs & rename currency levels
currency_1 = currency %>%
  pivot_longer(!ID, values_to = "HH.Income", names_to = "Currency", values_drop_na = TRUE) %>%
  mutate(
    HH.Income = fct_recode(
      HH.Income,
      `Prefer not to say` = "12"
    )) %>%
  mutate(HH.Income = as.factor(HH.Income))
```

```{r}
# drop rows with duplicated income info
id_to_del = currency_1 %>%
  group_by(ID) %>%
  count(ID, sort = TRUE) %>%
  filter(n > 1)

currency_2 = currency_1 %>%
  filter(!(ID %in% id_to_del$ID & Currency == "Other"))
```

```{r}
# merge with the main table
app_user_clean = left_join(app_user, currency_2)
head(app_user_clean, 10)
```

Here are the first 10 rows of the main table. We  keep NAs for now and will explore the missing values in the next section. For the levels of `HH.Income` column, authors of this survey already divided data into 12 intervals to address the problem of inconsistent income levels in different countries (1 for the lowest, 11 for the highest financial level and 12 for *"Prefer not to say"*). We therefore decide to not convert these currencies into a uniform unit and to go with the original income numbers. We will explain in detail in the results part.

```{r}
#export csv
write.csv(app_user_clean, "output/app_user_clean.csv", row.names = FALSE)
write.csv(app_type, "output/app_type.csv", row.names = FALSE)
write.csv(WhyPay, "output/WhyPay.csv", row.names = FALSE)
```
