==> 06-interactive.Rmd <==
# Interactive component

Our interactive graph uses D3 (version 6). Users can put mouse over the colored rectangles to access info of different App stores in each country. 



<script src="https://d3js.org/d3.v7.js"></script>
<script>
    <div class="row">
        <!-- Bootstrap Grid adapted to window size-->
        <div class="col-md-1 col-xs-1"></div>
        <div class="col-md-10 col-xs-10">
            <h2><b>App Store Distribution for Global Mobile Users</b></h2>
            <h5>**Put your mouse over the colored rectangles to access info of different App stores in each country</h5>
        </div>
        <div class="col-md-1 col-xs-1"></div>
    </div>

<meta charset="utf-8">
<script src="https://d3js.org/d3.v6.js"></script>
<div id="interactive"></div>

<script>

// margin & svg set up
const margin = {top: 10, right: 30, bottom: 70, left: 50},
    width = 460 - margin.left - margin.right,
    height = 400 - margin.top - margin.bottom;

const svg = d3.select("#interactive")
  .append("svg")
    .attr("width", width + margin.left + margin.right)
    .attr("height", height + margin.top + margin.bottom)
  .append("g")
    .attr("transform", `translate(${margin.left},${margin.top})`);
    
svg.append("text")
        .attr("x", (width / 2))             
        .attr("y", margin.top/2)
        .attr("text-anchor", "middle")  
        .style("font-size", "16px") 
        .style("text-decoration", "underline")  
        .text("App Store Distribution for Global Mobile Users");
  
// read CSV data
d3.csv("https://raw.githubusercontent.com/RuopuFan/MobileAppUsage/main/output/AppStore_Country.csv").then( function(data) {

  // read header & store country names
  const header = data.columns.slice(1)
  const country = data.map(d => d.Nationality)

  // Add X axis
  const x = d3.scaleBand()
      .domain(country)
      .range([0, width])
      .padding([0.2])
  svg.append("g")
    .attr("transform", `translate(0, ${height})`)
    .call(d3.axisBottom(x).tickSizeOuter(0))
  .selectAll("text")	
        .style("text-anchor", "end")
        .attr("dx", "-.8em")
        .attr("dy", ".15em")
        .attr("transform", "rotate(-45)");

  // Add Y axis
  const y = d3.scaleLinear()
    .domain([0, 600])
    .range([ height, 0 ]);
  svg.append("g")
    .call(d3.axisLeft(y));

  // set color palette
  const color = d3.scaleOrdinal()
    .domain(header)
    .range(d3.schemeTableau10);

  // stack the data
  const stackedData = d3.stack()
    .keys(header)
    (data)


  // create a pointer that shows actual data of each rectangle
  var pointer = d3.select("#interactive")
    .append("div")
    .style("opacity", 0)
    .attr("class", "tooltip")
    .style("background-color", "black")
    .style("color", "white")
  	.style("width", "150px")
  	.style("text-align", "center")
    .style("border-radius", "5px")
    .style("padding", "10px")

  // Show the bars
  svg.append("g")
    .selectAll("g")
    .data(stackedData)
    .join("g")
      .attr("fill", d => color(d.key))
      .attr("class", d => "myRect " + d.key )
      .selectAll("rect")
      .data(d => d)
      .join("rect")
        .attr("x", d => x(d.data.Nationality))
        .attr("y", d => y(d[1]))
        .attr("height", d => y(d[0]) - y(d[1]))
        .attr("width",x.bandwidth())
        .attr("stroke", "white")
        .on("mouseover", function (event,d) { 
    // actions to take when mouse over each rectangle
    // highlight app store across countries
          const countryName =d3.select(this.parentNode).datum().key;
    			d3.selectAll(".myRect").style("opacity", 0.3);  
          d3.selectAll("."+countryName).style("opacity",1);
    // show details about app store hovered
    			var userCount = d.data[countryName];
    			pointer.html("App Store: " + countryName + "<br>" + "User Count: " + userCount)
            .style("opacity", 1);
        })
        .on("mousemove", function(event,d){
    // actions to take when mouse moves
          pointer.style('transform', `translate(${event.layerX - 30}px, ${event.layerY - 300}px)`)
				})
        .on("mouseleave", function (event,d) { 
    // actions to take when the mouse leave the rectangle
          d3.selectAll(".myRect")
          .style("opacity",1);
    			pointer.style("opacity", 0);
      	})
})

</script>

[Data](https://github.com/RuopuFan/MobileAppUsage/blob/main/output/AppStore_Country.csv) and [codes](https://github.com/RuopuFan/MobileAppUsage/blob/main/d3-interactive.html) used in the graph are available in our repository.

References:

- Stacked Barplot with Group Highlighting: https://www.d3-graph-gallery.com/graph/barplot_stacked_highlight.html
