==> 07-conclusion.Rmd <==
# Conclusion

App Stores operate in a multisided market. With a rapidly increasing number of apps, software developers face the challenge of satisfying consumers' ever-rising and changing expectations. In this project, we narrow our focus on **Android** and **Apple iOS**, the two most used app store worldwide, and mainly investigate user price sensitivity and user income distribution for these two app stores across countries. We show that while owning an iPhone seems to be a strong indicator of higher income, this is more likely to be a geo-specific trend in North America given Apple's brand power, and Android app store gains dominance in countries such as India, South Korea, and Russia.

In the second part, we focus on analysis about **gender-neutral categories** and explore the distribution of user demographic information such as age, nationality, and years of education. While we do observe some cultural difference in downloads for gender-neutral categories, user persona in terms of age and educational background for gender-neutral apps are very similar in general.

Then we take a step further to investigate the **price-sensitive** users. The income distributions appear to be very different in each country or continent due to the random survey. In details, the visualization shows that there are a large number of student respondents in countries like Korea and Brazil, while incomes in Canada, United States and Great Britain are more normally distributed. We also use the word clouds of price-sensitive vs non-sensitive users to present reasons why people download apps that they spend money on. Price-sensitive users mainly recall their purchase experiences with **games** such as Angry Birds and Zombies as well as **pro** version of applications. If we look at the specific reasons why price-sensitive people pay for apps, the most popular choices are *"Can’t find a free app with similar features"* and *"App is on sale"*. As expected, sales promotion plays an important role in the ways that developers attract more users. Another significant finding is that many people spend on apps because they believe *"Paid apps have better quality/more features than free apps in general"*. Mobile app developers should start with improving their contents and features to meet consumers' expectation to stand out in such a competitive market.

Aside from our findings, there is still room to improve. Since the data is collected through a questionnaire, answers could be subjective to personal experience and cultural background. Besides, since almost all of the attributes we have are categorical variables, the type of data exploratory analysis we can apply is limited and the published dataset omits respondents' annual income, a key factor influencing user behavior, due to privacy concerns. Therefore, we can't build models such as regression models, which potentially enable us to get more insights of relationships between variables and predict future trends. Lastly, the survey collected data across 15 countries, and some questions collect text responses. At this point, we mainly focus on answers with hard-coded choice or are typed in English. In the future, we can incorporate techniques from NLP to extract additional information to get a more comprehensive picture.
